SouthPath and Fixes and UI stuff
Created the layout for South Path. Did not make any keys or lock any door.

Made a warp plate that spawns upon color key being picked up that sends player back to the hub.

Made a pause button on screen and implemented the pause UI.

Made adjustments and fixes to the black space and the camera.

Other adjustments/bug fixes.